# Make `findall` faster for AbstractArrays