@musm 2020-12-15
This  looks good to me. 
There's only one scenario where the only benefit of the current method is fewer allocations at the expense of a 2x slow-down in speed. That sounds like a good tradeoff to me. 

@mbauman @timholy   want to also sign off / review ?