@mbauman 2020-12-15
This is rather surprising to me — I suppose it's entirely due to the ability to sum logical arrays and `@inbounds` over a pre-determined size array.  I appreciate that this is better performance in many cases, but I dislike how this is harder to opt-out of than it is to opt-into.  I suppose, though, that — by definition — you're doing vectorized allocat-y Julia by using `findall`... and if this were truly chasing peak allocation-free performance, you'd do the iterative `find` loop.

So with that rationale, I'll give this a ✅.  I'll give it a bit more time, but barring further comments let's merge it tomorrow.