@garrison 2021-04-21
This change [broke UniqueVectors.jl](https://github.com/garrison/UniqueVectors.jl/issues/14).  My current plan is to issue a new release to fix it, but I am wondering: could there be other fallout throughout the package ecosystem?  Seems somewhat unlikely, but perhaps worth considering/checking.