@nalimilan 2021-04-21
Usually before releasing a new version all package tests are run against it to detect any problems in advance, so I guess we'll find out at that point.