@mbauman 2021-04-21
Would we avoid the ambiguity if it's `findall(testf::Function, A::Union{AbstractArray, Tuple})`?  Broadcasting for tuples is likely faster too.

While we try to avoid ambiguity changes like this, it's really hard to avoid and is easy to fix.