@garrison 2021-04-21
> Would we avoid the ambiguity if it's `findall(testf::Function, A::Union{AbstractArray, Tuple})`?

Do you mean if we were to replace the signature `findall(testf::Function, A::AbstractArray)` with `findall(testf::Function, A::Union{AbstractArray, Tuple})`?  The method signature in UniqueVectors that is now ambiguous (following the change in this pull request) is `findall(p::Base.Fix2{typeof(in),<:AbstractUniqueVector}, a::Union{Tuple, AbstractArray})`.  I believe that even if the change you are suggesting is made, the UniqueVector method would still be ambiguous with `findall(pred::Fix2{typeof(in)}, x::AbstractArray)` and `findall(pred::Fix2{typeof(in)}, x::Tuple)`.

An alternative approach (and one which you might actually be implying) is to replace the three signatures of `findall` introduced in this PR with a single signature: the one you mentioned, `findall(testf::Function, A::Union{AbstractArray, Tuple})`.  Then this method could e.g. call an inner `_findall` method, and the dispatching could be performed one level deeper.  I *think* this would result in unambiguous methods, but haven't considered it carefully or tested it.

> While we try to avoid ambiguity changes like this, it's really hard to avoid and is easy to fix.

Do you mean to fix it by adjusting julia before the release or to fix it in packages? I am on board either way.  If there is no other fallout in the package ecosystem, then the above solution is almost certainly overkill -- the UniqueVectors package is probably the right place to fix it.