@mbauman 2021-04-21
Yes, I meant the latter suggestion on both counts. Sorry for the brevity!